requirements.txt
feedparser>=6.0.0
anthropic>=0.40.0
httpx>=0.27.0
python-dateutil>=2.8.0
schedule>=1.2.0
